This push brings more improvements to the release candidate for the new pants build system as well as a good start towards full py 3.x compatibility.
    pants:
      + Get depmap graphviz support working on dotty - omnigraffle was hiding a few bugs.
      + Add support for passing tests to the junit runner and the specs runner via @argfiles for command lines that are beyond a size threshold.
      + Divorce the ide algorithm from bang so it can operate on unchunked targets and avoid cycles that can result in aggregated ide targets.
      + Add a markdown_to_html task.
      + Make javac and scalac tasks resilient to targets with no files.
      + Fix jar_publish change tracking to include internal and external dependency changes.
      + Fix codegen tasks to always link synthetic targets for upstream even when actually skipping codegen.
      + Move mostly to __future__ print_function make pants bootstrappable in python 3.x add pants.multi to manage pants with multiple python interpreters/
      + Harden BuildFile against valid BUILD file paths that are in fact dirs.
      + Add Phase.uninstall to support trimming away un-needed pants goals.
      + Fixup target ids to handle a top-level BUILD and compiler class mapping to normalize path names for comparison.
      + Change the 1-arg FuncTask wrapper to take a context - this will be a much more common use-case for func actions.
      + Add deploy_excludes to jvm_binary and respect these in the binary and bundle tasks as a short-term hack to be replaced with proper ivy configuration/maven scope support.
      + Add support for globs and rglobs when defining bundles

    python:
      + Make python twitter.commons more py3x friendly.
      + Changes to python common log:
        + change logs from going to stdout to going to stderr
        + change the default log level for stderr to ERROR rather than NONE
        + if we are writing both to disk and to stderr, write a message stating where we write the log files

    jvm:
      Upgrade to guava 11.0.2, twitter-text 1.4.8 and finagle/util 2.0.0.

      app:
        + Log more exception info in applauncher.

      args:
        + Refactoring of command line args parsing.
        + Scala prototype command line parsing implementation.

+++ src/python/twitter/common/quantity/__init__.py
@@ -19,9 +19,9 @@ from numbers import Integral
 class AmountUnit(object):
   def __init__(self, multiplier, base_or_unit, display):
     if isinstance(base_or_unit, Integral):
-      self._multiplier = multiplier * long(base_or_unit)
+      self._multiplier = multiplier * int(base_or_unit)
     elif isinstance(base_or_unit, AmountUnit):
-      self._multiplier = multiplier * long(base_or_unit.multiplier())
+      self._multiplier = multiplier * int(base_or_unit.multiplier())
     else:
       raise TypeError('Unknown type for base_or_unit: %s' % type(base_or_unit))
     self._display = display
